requirements.txt
betterproto~=1.2.5
discord.py~=2.0.1
cachetools~=5.2.0
loguru~=0.6.0
rtoml~=0.8.0
pydantic~=1.10.2
aiohttp~=3.8.1